Change lisp_string_to_c_string length parameter to be size_t
The argument passed to the length parameter of lisp_string_to_c_string is
always the sizeof an array, and since it is compared to the Lisp unsigned
length of the Lisp string is more appropriately a size_t than an int.
Likewise for the index into a string.

+++ inc/uutilsdefs.h
@@ -1,6 +1,6 @@
 #ifndef UUTILSDEFS_H
 #define UUTILSDEFS_H 1
-int lisp_string_to_c_string(LispPTR Lisp, char *C, int length);
+int lisp_string_to_c_string(LispPTR Lisp, char *C, size_t length);
 int c_string_to_lisp_string(char *C, LispPTR Lisp);
 LispPTR check_unix_password(LispPTR *args);
 LispPTR unix_username(LispPTR *args);
